backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
PyYAML
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
pytest-asyncio
